mobile fixes

+++ src/main/webapp/WEB-INF/jsp/assetBrowser.jsp
@@ -284,7 +284,10 @@
           var currentPage = gridOptions.api.paginationGetCurrentPage();
           gridOptions.api.paginationGoToPage(currentPage-1);
         } else if (event.target.id == "delete") {
-          delete_selected();
+          var r = confirm("Delete the selected asset?");
+          if (r) {
+            delete_selected();
+          }
         }
         updateGridData({});
       }
@@ -303,10 +306,7 @@
       if (!adminLoggedIn) {
         // Disable the user browser link in the navbar if the logged in
         // user does not have admin access
-        document.getElementById("userBrowser").setAttribute('class', 'nav-item disabled');
-        document.getElementById("userBrowserLink").click(function() {
-          return false;
-        });
+        document.getElementById("userBrowserLink").href = "#";
       }
 
       // Execute an HTTP call to get the available asset metadata
